feat(ace): slice 8.2 T2 — extract packageHash to package-hash.ts, sweep 9 import sites
Remove the OLD signature-INCLUDING packageHash from resolve.ts (drop now-unused
createHash + canonicalBytes imports) and point all 9 import sites at the new
signature-EXCLUDING ./package-hash.ts (T1). solver*.test.ts switch to import type
{ FetchPackage } (verbatimModuleSyntax). package_hash VALUES change (signature
excluded) — intentional, pre-release.

One test-fixture update (NOT a behavioral regression):
ace.test.ts '--frozen with a malformed ... locked node refuses cleanly' used
packageHash({}) as a setup helper to derive a pin. The new packageHash destructures
pkg.manifest, so it throws on a manifest-less {} payload (vs the old leniency). The
frozen-install path's PASS-1 shape guard (ace.ts:986-989) refuses the malformed node
BEFORE packageHash is ever called, so the pin value is never compared — verified the
install still refuses cleanly (exit 1, no throw) with a placeholder pin. Replaced the
packageHash({}) helper with a placeholder hash; the test's intent + assertions are
unchanged. No literal-hash pins needed updating.

tsc 0, bun test tools/ace/ 364 pass 0 fail.
