+++ tools/ace/registry-publish.ts
@@ -2,7 +2,7 @@
 // Inverts the consumer: build the signed IndexSignableContent a slice-6 consumer accepts.
 // No I/O — the caller (ace.ts) reads the package files + sequence + pem and writes the output.
 import type { AcePackage, RegistryEntry } from "./store.ts";
-import { packageHash } from "./resolve.ts";
+import { packageHash } from "./package-hash.ts";
 import type { IndexSignableContent, RevocationMap } from "./signing.ts";
 import { signIndex } from "./signing.ts";
 import type { IndexDoc } from "./registry-remote.ts";
