+++ tools/ace/solver.test.ts
@@ -1,6 +1,7 @@
 import { describe, expect, test } from "bun:test";
 import { solve } from "./solver.ts";
-import { packageHash, type FetchPackage } from "./resolve.ts";
+import type { FetchPackage } from "./resolve.ts";
+import { packageHash } from "./package-hash.ts";
 import { contentHash } from "./store.ts";
 import type { AcePackage, AceDependency, RegistryEntry } from "./store.ts";
